we dont even need to decorate on the obj
